requirements.txt
gunicorn==20.1.0
whitenoise==5.3.0

django==3.2.7

redis==3.5.3
psycopg2==2.9.1
mysqlclient==2.0.1
dj_database_url
django-ordered-model==3.4.3

django-ninja==0.14.0
pydantic-django==0.1.1
huey==2.3.1

django-storages==1.11.1

django-bootstrap4==3.0.1

vk_api==11.9.4
googletrans==3.1.0a0
boto3==1.18.42
pyTelegramBotAPI==4.0.0

pythonanywhere
Pillow
beautifulsoup4==4.9.1
requests==2.25.1
